+++ src/fundamend/sqlmodels/anwendungshandbuch.py
@@ -523,69 +523,4 @@ class Anwendungshandbuch(SQLModel, table=True):
         )
 
 
-class AhbHierarchyMaterialized(SQLModel, table=True):
-    """
-    A materialized flattened AHB hierarchy containing segment groups, segments, data elements, codes,
-    and enriched with metadata like format, versionsnummer, and prüfidentifikator.
-    This table is not thought to be written to, but only read from.
-    It is created once after all other tables have been filled by the create_ahb_view function in ahbview.py.
-    """
-
-    __tablename__ = "ahb_hierarchy_materialized"
-    id: UUID = Field(default_factory=uuid.uuid4, primary_key=True)
-    anwendungsfall_pk: UUID = Field(index=True)
-    current_id: UUID
-    root_id: UUID
-    parent_id: Optional[UUID] = None
-    depth: int
-    position: Optional[int] = Field(default=None)
-    path: str
-    id_path: str
-    parent_path: str
-    root_order: int
-    type: str = Field(index=True)
-    source_id: UUID
-    sort_path: str = Field(index=True)
-
-    # Metadata
-    pruefidentifikator: str = Field(index=True)
-    format: str = Field(index=True)
-    versionsnummer: str = Field(index=True)
-    gueltig_von: Optional[date] = Field(default=None, index=True)
-    gueltig_bis: Optional[date] = Field(default=None, index=True)
-    kommunikation_von: Optional[str] = Field(default=None, index=True)
-    beschreibung: Optional[str] = Field(default=None, index=True)
-    edifact_format_version: Optional[EdifactFormatVersion] = Field(default=None, index=True)
-
-    # Segment Group
-    segmentgroup_id: Optional[str] = Field(default=None, index=True)
-    segmentgroup_name: Optional[str] = Field(default=None, index=True)
-    segmentgroup_ahb_status: Optional[str] = Field(default=None)
-    segmentgroup_position: Optional[int] = Field(default=None, index=True)
-    segmentgroup_anwendungsfall_primary_key: Optional[UUID] = Field(default=None)
-
-    # Segment
-    segment_id: Optional[str] = Field(default=None, index=True)
-    segment_name: Optional[str] = Field(default=None, index=True)
-    segment_number: Optional[str] = Field(default=None, index=True)
-    segment_ahb_status: Optional[str] = Field(default=None)
-    segment_position: Optional[int] = Field(default=None, index=True)
-
-    # Data Element Group
-    dataelementgroup_id: Optional[str] = Field(default=None, index=True)
-    dataelementgroup_name: Optional[str] = Field(default=None, index=True)
-    dataelementgroup_position: Optional[int] = Field(default=None, index=True)
-
-    # Data Element
-    dataelement_id: Optional[str] = Field(default=None, index=True)
-    dataelement_name: Optional[str] = Field(default=None, index=True)
-    dataelement_position: Optional[int] = Field(default=None, index=True)
-    dataelement_ahb_status: Optional[str] = Field(default=None, index=True)
-
-    # Code
-    code_id: Optional[UUID] = Field(default=None, index=True)
-    code_name: Optional[str] = Field(default=None, index=True)
-    code_description: Optional[str] = Field(default=None, index=True)
-    code_value: Optional[str] = Field(default=None, index=True)
-    code_ahb_status: Optional[str] = Field(default=None, index=True)
-    code_position: Optional[int] = Field(default=None, index=True)
+__all__ = ["Code", "DataElement", "DataElementGroup", "Segment", "SegmentGroup", "Anwendungsfall", "Bedingung"]
